var/www/example/requirements.txt
Flask>=3.0.2
pyopenssl>=23.2.0
dnspython>=2.6.1
certifi>=2023.11.17
bleach>=6.1.0
python-whois>=0.9.4
python-telegram-bot>=21.3
prettytable>=3.10.0
ipwhois>=1.1.0
urllib3>=2.0.7
json2html>=1.3.0